+++ integration_tests/suite/base/test_users.py
@@ -1055,8 +1055,15 @@ def generate_user_resources_bodies(
 @fixtures.switchboard()
 @fixtures.device()
 @fixtures.user()
+@fixtures.group()
 def test_post_update_delete_full_user_no_error(
-    group_extension, group, funckey_template, switchboard, device, user_destination
+    group_extension,
+    group,
+    funckey_template,
+    switchboard,
+    device,
+    user_destination,
+    new_group,
 ):
     (
         exten,
@@ -1273,7 +1280,12 @@ def test_post_update_delete_full_user_no_error(
         }
 
         response = url.put(
-            {**user_body, 'fallbacks': {**new_fallbacks}, 'forwards': {**new_forwards}},
+            {
+                **user_body,
+                'fallbacks': {**new_fallbacks},
+                'forwards': {**new_forwards},
+                'groups': [{'uuid': new_group['uuid']}],
+            },
             query_string="recursive=True",
         )
         response.assert_updated()
@@ -1295,7 +1307,15 @@ def test_post_update_delete_full_user_no_error(
             has_entries(**new_forwards),
         )
 
-        # retrieve the data for the user and check the data returned (fallbacks and forwards)
+        # retrieve the groups for the user and check the data
+        assert_that(
+            confd.groups(new_group['uuid']).get().item,
+            has_entries(
+                members=has_entries(users=contains(has_entries(uuid=payload['uuid'])))
+            ),
+        )
+
+        # retrieve the data for the user and check the data returned (fallbacks, forwards and groups)
         assert_that(
             confd.users(payload['uuid']).get().item,
             has_entries(
@@ -1306,6 +1326,13 @@ def test_post_update_delete_full_user_no_error(
                     fail_destination=has_entries(**destination),
                 ),
                 forwards=has_entries(**new_forwards),
+                groups=contains(
+                    has_entries(
+                        id=new_group['id'],
+                        uuid=new_group['uuid'],
+                        name=new_group['name'],
+                    )
+                ),
             ),
         )